fix(ci): restore PR checks — align recovery fixtures and structural gates
