+++ tests/unit/cli/test_service_cli_parts/test_service_cli_part_003.py
@@ -504,12 +504,16 @@ def test_worker_entrypoint_wires_control_worker_dependencies(
             stack_launcher: object,
             config: object,
             service_diagnostics: object = None,
+            before_provision: object = None,
+            after_provision: object = None,
         ) -> None:
             created["provisioner_session_factory"] = session_factory
             created["provisioner_git"] = git
             created["provisioner_stack_launcher"] = stack_launcher
             created["provisioner_config"] = config
             created["provisioner_service_diagnostics"] = service_diagnostics
+            created["provisioner_before_provision"] = before_provision
+            created["provisioner_after_provision"] = after_provision
 
     class _ControlWorker:
         def __init__(
@@ -618,6 +622,8 @@ def test_worker_entrypoint_wires_control_worker_dependencies(
     assert created["worker_claude_base_reaper"] is not None
     assert created["worker_terminal_gc_reaper"] is not None
     assert created["provisioner_config"].node_id == "node-1"
+    assert callable(created["provisioner_before_provision"])
+    assert callable(created["provisioner_after_provision"])
     assert created["worker_config"].poll_interval_seconds == 0.25
     assert created["worker_config"].max_concurrent_provisions == 2
     assert created["worker_config"].max_concurrent_executions == 4
